src/github-app-preview-url/pyproject.toml
[project]
name = "github-app-preview-url"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.39.0",
]
